// types.ts
export enum UserRole {
  ADMIN = 'admin',
  COMPANY = 'company',
  EMPLOYEE = 'employee',
  GROUP = 'group',
}

export enum UserStatus {
  PENDING = 'pending',
  APPROVED = 'approved',
  REJECTED = 'rejected',
}